tmp4

+++ temperature_conv/src/lib.rs
@@ -16,7 +16,7 @@ Formula
 */
 
 pub fn celsius_to_fahrenheit(c: f64) -> f64 {
-    to_fixed2((c*((9/5) as f64)) + (32 as f64))
+    to_fixed2( c * (9.0/5.0) + 32.0 )
 }
 
 #[cfg(test)]
